ИИ Ассистент/requirements.txt
python-telegram-bot>=22.5
openai>=2.15.0
chromadb>=1.4.1
sentence-transformers==2.3.1
gspread==5.12.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
requests==2.31.0
tqdm>=4.65.0
